pyproject.toml
[project]
name = "cqrs-course-dataset"
version = "0.1.0"
description = ""
authors = [
    {name = "hagyutae",email = "gted221@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "openai (>=1.99.5,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
